pyproject.toml
[project]
name = "ultragent"
version = "0.1.0"
description = "Sistema de Ingeniería Híbrida Autónoma - MCP Server"
readme = "README.md"
requires-python = ">=3.12, <3.13"
license = { text = "MIT" }
authors = [
    { name = "Ultragent Architect" }
]
keywords = ["mcp", "ai", "agent", "automation"]

dependencies = [
    "chromadb>=1.4.1",
    "docker>=7.1.0",
    "fastmcp>=2.0.0",
    "httpx>=0.28.1",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=5.2.2",
    "tree-sitter==0.24.0",
    "tree-sitter-javascript==0.23.0",
    "tree-sitter-python==0.23.2",
    "watchdog>=4.0.0",
    "duckduckgo-search>=5.0.0",
    "rank-bm25>=0.2.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["mcp_server.py", "setup_fs.py"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
